requirements.txt
yfinance==0.2.52
pandas==2.2.3
numpy==2.1.2
streamlit==1.39.0
requests==2.32.3
python-dotenv==1.0.1
